feat(insulin-ww-ratio-suggestion): baseDose float→int conversion (p1)
Converts PatientProfile::$baseDose and DiaryEntry::$baseDoseSnapshot
from float to int, required so the upcoming base-dose suggestion
algorithm can produce whole-unit values.

- src/Entity/PatientProfile.php, src/Entity/DiaryEntry.php: type change
- src/Form/ProfileFormType.php: baseDose field switches to IntegerType
- migrations/Version20260826073658.php: Postgres column type migration
  (USING ROUND(...)::INTEGER cast), applied to dev + test DBs
- Existing test literals updated across PatientProfileTest,
  DiaryEntryTest, DiaryControllerTest, ProfileControllerTest,
  OnboardingControllerTest, HomeControllerTest, SecurityControllerTest
- context/foundation/roadmap.md: S-03 flipped to in-progress

+++ src/Entity/DiaryEntry.php
@@ -48,12 +48,12 @@ class DiaryEntry
     private float $insulinWwRatioSnapshot;
 
     #[ORM\Column]
-    private float $baseDoseSnapshot;
+    private int $baseDoseSnapshot;
 
     #[ORM\Column]
     private \DateTimeImmutable $createdAt;
 
-    public function __construct(User $user, int $glycemiaMgDl, \DateTimeImmutable $measuredAt, float $insulinWwRatioSnapshot, float $baseDoseSnapshot)
+    public function __construct(User $user, int $glycemiaMgDl, \DateTimeImmutable $measuredAt, float $insulinWwRatioSnapshot, int $baseDoseSnapshot)
     {
         $this->user = $user;
         $this->glycemiaMgDl = $glycemiaMgDl;
@@ -150,7 +150,7 @@ class DiaryEntry
         return $this->insulinWwRatioSnapshot;
     }
 
-    public function getBaseDoseSnapshot(): float
+    public function getBaseDoseSnapshot(): int
     {
         return $this->baseDoseSnapshot;
     }
